# macho: faster and more memory efficient linker