… CMakeLists.txt
… src/link/MachO/Atom.zig
… src/link/MachO/Relocation.zig
… src/link/MachO/Trie.zig
… src/link/MachO/ZldAtom.zig
… src/link/MachO/dead_strip.zig
… src/link/MachO/thunks.zig
… src/link/MachO/zld.zig