@andrewrk 2022-10-22
You can count the drone CI as a success. It passed all the tests and simply ran out of CI time when it was creating the tarball at the end. Furthermore, this is affecting master branch and 10b8c4d8f02c085090b07f3709de2d70d3b85cf0 should reduce failure rate.