@kubkon 2022-10-22
> What a beauty.
> 
> btw you can give 3 commands at the same time to hyperfine

Oh shit, did not know that! I'll keep that in mind for the future, thanks!